+++ frontend/src/components/ui/label.tsx
@@ -1,7 +1,7 @@
 import * as LabelPrimitive from "@radix-ui/react-label"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/tw-merge.utils"
 
 function Label({
     className,
