+++ frontend/src/components/ui/select.tsx
@@ -4,7 +4,7 @@ import {
 } from "lucide-react"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/tw-merge.utils"
 
 function Select({
     ...props
